Mise à jour de la gestion des avatars

+++ srcs/app/transcendence/methods/edit_profile.py
@@ -46,9 +46,12 @@ def edit_profile(request):
 			return JsonResponse({'status': 'error', 'message': 'Invalid birthdate.'}, status=400)
 
 	if 'avatar' in request.FILES:
+		avatar_file = request.FILES['avatar']
+		if avatar_file.size > 1024 * 1024:
+			return JsonResponse({'status': 'error', 'message': 'Image is too large. (Max 1MB)'}, status=400)
 		if user.avatar:
 			user.avatar.delete()
-		user.avatar = request.FILES['avatar']
+		user.avatar = avatar_file
 
 	password = data.get('password')
 	password_confirm = data.get('password_confirm')
